Navigation changes and select for SizeType in product creation

+++ src/Shoppu.Application/Products/Commands/CreateProduct/CreateProductCommand.cs
@@ -45,7 +45,7 @@ namespace Shoppu.Application.Products.Commands.CreateProduct
                 Code = request.ProductViewModel.Code,
                 BaseSlug = baseSlug,
                 IsAccessible = false,
-                SizeTypeId = request.ProductViewModel.SizeTypeId
+                SizeTypeId = (int)request.ProductViewModel.SizeTypeId
             };
             await _context.Products.AddAsync(product);
             await _context.SaveChangesAsync(cancellationToken);
